+++ src/layers/XRLayer/xr-layer-fragment-colormap.webgl2.glsl
@@ -13,7 +13,7 @@ uniform SAMPLER_TYPE channel4;
 uniform SAMPLER_TYPE channel5;
 
 // range
-uniform vec2 sliderValues[6];
+uniform vec2 contrastLimits[6];
 
 // opacity
 uniform float opacity;
@@ -28,12 +28,12 @@ out vec4 color;
 
 
 void main() {
-  float intensityValue0 = sample_and_apply_sliders(channel0, vTexCoord, sliderValues[0]);
-  float intensityValue1 = sample_and_apply_sliders(channel1, vTexCoord, sliderValues[1]);
-  float intensityValue2 = sample_and_apply_sliders(channel2, vTexCoord, sliderValues[2]);
-  float intensityValue3 = sample_and_apply_sliders(channel3, vTexCoord, sliderValues[3]);
-  float intensityValue4 = sample_and_apply_sliders(channel4, vTexCoord, sliderValues[4]);
-  float intensityValue5 = sample_and_apply_sliders(channel5, vTexCoord, sliderValues[5]);
+  float intensityValue0 = sample_and_apply_contrast_limits(channel0, vTexCoord, contrastLimits[0]);
+  float intensityValue1 = sample_and_apply_contrast_limits(channel1, vTexCoord, contrastLimits[1]);
+  float intensityValue2 = sample_and_apply_contrast_limits(channel2, vTexCoord, contrastLimits[2]);
+  float intensityValue3 = sample_and_apply_contrast_limits(channel3, vTexCoord, contrastLimits[3]);
+  float intensityValue4 = sample_and_apply_contrast_limits(channel4, vTexCoord, contrastLimits[4]);
+  float intensityValue5 = sample_and_apply_contrast_limits(channel5, vTexCoord, contrastLimits[5]);
 
   float intensityArray[6] = float[6](intensityValue0, intensityValue1, intensityValue2, intensityValue3, intensityValue4, intensityValue5);
   float intensityCombo = 0.;