+++ tests/layers_views/layers/utils.spec.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-extraneous-dependencies, no-unused-expressions */
 import test from 'tape-catch';
 import { range } from '../../../src/layers/MultiscaleImageLayer/utils';
-import { padWithDefault, padColorsAndSliders } from '../../../src/layers/utils';
+import { padWithDefault, padColorsAndWindows } from '../../../src/layers/utils';
 
 test('range test', t => {
   const expected = [0, 1, 2];
@@ -20,13 +20,13 @@ test('padWithDefault test', t => {
   t.end();
 });
 
-test('padColorsAndSliders test', t => {
+test('padColorsAndWindows test', t => {
   const expectedNoDomain = {
-    paddedSliderValues: [0, 5, 0, 5, 255, 255, 255, 255, 255, 255, 255, 255],
-    paddedColorValues: [1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
+    paddedContrastLimits: [0, 5, 0, 5, 255, 255, 255, 255, 255, 255, 255, 255],
+    paddedColors: [1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
   };
   const expectedChannelOff = {
-    paddedSliderValues: [
+    paddedContrastLimits: [
       0,
       5,
       255,
@@ -40,10 +40,10 @@ test('padColorsAndSliders test', t => {
       255,
       255
     ],
-    paddedColorValues: [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
+    paddedColors: [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
   };
   const expectedDomain = {
-    paddedSliderValues: [
+    paddedContrastLimits: [
       0,
       5,
       0,
@@ -57,10 +57,10 @@ test('padColorsAndSliders test', t => {
       1000,
       1000
     ],
-    paddedColorValues: [1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
+    paddedColors: [1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
   };
   const expected16Bit = {
-    paddedSliderValues: [
+    paddedContrastLimits: [
       0,
       5,
       0,
@@ -74,52 +74,52 @@ test('padColorsAndSliders test', t => {
       2 ** 16 - 1,
       2 ** 16 - 1
     ],
-    paddedColorValues: [1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
+    paddedColors: [1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
   };
 
   t.deepEqual(
-    padColorsAndSliders({
-      sliderValues: [
+    padColorsAndWindows({
+      contrastLimits: [
         [0, 5],
         [0, 5]
       ],
-      colorValues: [
+      colors: [
         [255, 0, 0],
         [255, 255, 0]
       ],
-      channelIsOn: [true, true],
+      channelsVisible: [true, true],
       dtype: 'Uint8'
     }),
     expectedNoDomain,
     'Pads with no domain provided'
   );
   t.deepEqual(
-    padColorsAndSliders({
-      sliderValues: [
+    padColorsAndWindows({
+      contrastLimits: [
         [0, 5],
         [0, 5]
       ],
-      colorValues: [
+      colors: [
         [255, 0, 0],
         [255, 255, 0]
       ],
-      channelIsOn: [true, false],
+      channelsVisible: [true, false],
       dtype: 'Uint8'
     }),
     expectedChannelOff,
     'Pads with one channel turned off'
   );
   t.deepEqual(
-    padColorsAndSliders({
-      sliderValues: [
+    padColorsAndWindows({
+      contrastLimits: [
         [0, 5],
         [0, 5]
       ],
-      colorValues: [
+      colors: [
         [255, 0, 0],
         [255, 255, 0]
       ],
-      channelIsOn: [true, true],
+      channelsVisible: [true, true],
       domain: [0, 1000],
       dtype: 'Uint8'
     }),
@@ -127,16 +127,16 @@ test('padColorsAndSliders test', t => {
     'Pads with provided domain value'
   );
   t.deepEqual(
-    padColorsAndSliders({
-      sliderValues: [
+    padColorsAndWindows({
+      contrastLimits: [
         [0, 5],
         [0, 5]
       ],
-      colorValues: [
+      colors: [
         [255, 0, 0],
         [255, 255, 0]
       ],
-      channelIsOn: [true, true],
+      channelsVisible: [true, true],
       dtype: 'Uint16'
     }),
     expected16Bit,